Refactor logger to support context-based logging
- Added context parameter to all logging methods
- Simplified logger implementation by removing configuration methods
- Maintained existing logging levels and functionality
- Improved logger interface for better context tracking
